requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
easygui==0.98.1
Flask==0.12.2
get==2019.4.13
gunicorn==20.0.4
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.18.4
scipy==1.3.3
six==1.13.0
urllib3==1.22
Werkzeug==0.16.0
